fix: Move SpeakerWithConversationCountDTO to domain layer
Fix CRITICAL Clean Architecture violation where domain layer
(speaker_repository.py) was importing a DTO from application layer.

Changes:
- Create src/domain/dtos/ directory for domain-layer DTOs
- Move SpeakerWithConversationCountDTO to domain layer
- Update imports in speaker_repository.py to use domain DTO
- Update imports in speaker_repository_impl.py
- Add re-export in application layer for backward compatibility

This ensures proper dependency direction:
Application → Domain (correct)
Not: Domain → Application (violation)
