+++ src/domain/dtos/__init__.py
@@ -1,0 +1,1 @@
+"""Domain DTOs for repository contracts."""
